♻️ Restructure grid layout examples (core + extended)
