+++ apps/cookbook/src/app/examples/grid-layout-example/grid-layout-core-example/grid-layout-core-example.component.html
@@ -1,0 +1,144 @@
+<main class="debug">
+  <section>
+    <div class="safe-area-inline">
+      <div class="max-width-container gutter">
+        <h1>Værdier i alt</h1>
+      </div>
+      <div class="max-width-container">
+        <div class="grid-container">
+          <div class="grid-item half@tablet-up">
+            <article>
+              <div class="gutter">
+                <h2>Bolig</h2>
+              </div>
+              <kirby-card hasPadding="true">
+                <div class="box example-text align-center">
+                  1
+                </div>
+              </kirby-card>
+            </article>
+          </div>
+          <div class="grid-item half@tablet-up">
+            <article>
+              <div class="gutter">
+                <h2>Konti</h2>
+              </div>
+              <kirby-card hasPadding="true">
+                <div class="box example-text align-center">
+                  2
+                </div>
+              </kirby-card>
+            </article>
+          </div>
+          <div class="grid-item half@tablet-up">
+            <article>
+              <div class="gutter">
+                <h2>Pension</h2>
+              </div>
+              <kirby-card hasPadding="true">
+                <div class="box example-text align-center">
+                  3
+                </div>
+              </kirby-card>
+            </article>
+          </div>
+          <div class="grid-item half@tablet-up">
+            <article>
+              <div class="gutter">
+                <h2>Investering</h2>
+              </div>
+              <kirby-card hasPadding="true">
+                <div class="box example-text align-center">
+                  4
+                </div>
+              </kirby-card>
+            </article>
+          </div>
+        </div>
+      </div>
+    </div>
+  </section>
+
+  <div class="safe-area-inline">
+    <div class="max-width-container gutter">
+      <div class="box">
+        <button kirby-button size="lg">Tilføj</button>
+        <p class="align-center">Tilføj gæld og formue fra andre banker m.m.</p>
+      </div>
+    </div>
+  </div>
+
+  <section>
+    <div class="safe-area-inline">
+      <div class="max-width-container gutter">
+        <h1>Anbefalinger</h1>
+        <p>
+          Du har 3 nye anbefalinger til at forbedre din økonomi. Kontakt din rådgiver for at høre
+          mere.
+        </p>
+      </div>
+      <div class="max-width-container">
+        <div class="grid-container">
+          <div class="grid-item half@tablet-up third@desktop-up">
+            <article>
+              <kirby-card hasPadding="true">
+                <h2>Få mere ud af din opsparing</h2>
+                <p>
+                  Fluctuate interest rates Dow Jones receive rise government term municipal market
+                  Nikkei passively return performance. Public finance holder fiat established bonds
+                  hedge fund benchmark.
+                </p>
+              </kirby-card>
+            </article>
+          </div>
+          <div class="grid-item half@tablet-up third@desktop-up">
+            <article>
+              <kirby-card hasPadding="true">
+                <h2>Udnyt friværdi i din bolig</h2>
+                <p>
+                  Fluctuate interest rates Dow Jones receive rise government term municipal market
+                  Nikkei passively return performance. Public finance holder fiat established bonds
+                  hedge fund benchmark.
+                </p>
+              </kirby-card>
+            </article>
+          </div>
+          <div class="grid-item half@tablet-up third@desktop-up">
+            <article>
+              <kirby-card hasPadding="true">
+                <h2>Forøg din formue</h2>
+                <p>
+                  Fluctuate interest rates Dow Jones receive rise government term municipal market
+                  Nikkei passively return performance. Public finance holder fiat established bonds
+                  hedge fund benchmark.
+                </p>
+              </kirby-card>
+            </article>
+          </div>
+        </div>
+      </div>
+    </div>
+  </section>
+
+  <div class="not-grid">
+    <div class="safe-area-inline">
+      <div class="max-width-container gutter">
+        <div id="example-1" class="box example-text align-center">
+          <div class="h1">
+            max-width + gutter + safe area
+          </div>
+          .safe-area-inline > .max-width-container.gutter
+        </div>
+      </div>
+    </div>
+    <div class="safe-area-inline">
+      <div id="example-2" class="box example-text align-center">
+        <div class="h1">Full width + safe area</div>
+        .safe-area-inline
+      </div>
+    </div>
+    <div id="example-3" class="box example-text align-center">
+      <div class="h1">Full width</div>
+    </div>
+  </div>
+</main>
