+++ apps/cookbook/src/app/examples/grid-layout-example/grid-layout-core-example/grid-layout-core-example.component.scss
