+++ apps/cookbook/src/app/examples/grid-layout-example/grid-layout-core-example/grid-layout-core-example.component.ts
@@ -1,0 +1,8 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'cookbook-grid-layout-core-example',
+  templateUrl: './grid-layout-core-example.component.html',
+  styleUrls: ['./grid-layout-core-example.component.scss'],
+})
+export class GridLayoutCoreExampleComponent {}
